Fixed multiple integer overflow issues (not bugs, just intentional integer overflow; does not affect regression). Fixed stack overflow bug in the sample code.

+++ src/libFRFXLL/src/algorithm/delay.h
@@ -88,7 +88,7 @@ namespace FingerJetFxOSE {
         delay(size_t dt, bool init_val = false) 
           : ptr(buffer)
           , end(buffer + (dt + 7) / 8)
-          , initmask(1 << (-int8(dt) & 7))
+          , initmask(1 << (-dt & 7)) 
         {
           mask = initmask;
           while(ptr < end) {
